refactor: refactor adminhome into smaller jsx components
